Add full-only-mode using a queue fixes #18
A fixed-length queue, every time we get requested to send a full update
append the current time to this queue.
If that queue is full **and** the average of all values on this queue is
within a threshold of the current time, switch to full-only-mode for a
fixed amount of time.

Tuneables:
- queue length, i.e. how many of the newest requests do we consider,
depends on the number of neighbours
- trigger threshold, i.e. when to trigger full-only mode, depends on the
message interval and the number of neighbours
- full-only time, how long should we stay in full-only mode